refactor(strip): merge validate_annotations into strip_artifacts
